# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen API Documentation",
        "description": "Lumen API documentation generated using Swagger",
        "contact": {
            "email": "devahmad92@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/airports/{id}": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get a single airport",
                "operationId": "0a5ed466139705fda0b268eb5c90b769",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Airport ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Airport"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID format"
                    },
                    "404": {
                        "description": "Airport not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Airports"
                ],
                "summary": "Update an existing airport",
                "operationId": "51a6833cf435af4fa46a52085ed89111",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Airport ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass airport updated data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AirportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Airport updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Airport"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Airport not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Airports"
                ],
                "summary": "Delete an airport",
                "operationId": "9fbd7b87849cb6ff3c475e1897461b30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Airport ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Airport deleted"
                    },
                    "404": {
                        "description": "Airport not found"
                    }
                }
            }
        },
        "/airports": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "List all airports",
                "operationId": "e20ac2c0012afc0c6670c25213698380",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items to return per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Airport"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Airports"
                ],
                "summary": "Add a new airport",
                "operationId": "888836d22de80f20391b4b97a2acf631",
                "requestBody": {
                    "description": "Pass airport data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AirportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Airport created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Airport"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    }
}